+++ tiledb/query_condition.cc
@@ -28,19 +28,6 @@ private:
   Context ctx_;
   shared_ptr<QueryCondition> qc_;
 
-private:
-  PyQueryCondition(shared_ptr<QueryCondition> qc, tiledb_ctx_t *c_ctx)
-      : qc_(qc) {
-    ctx_ = Context(c_ctx, false);
-  }
-
-  void set_ctx(py::object ctx) {
-    tiledb_ctx_t *c_ctx;
-    if ((c_ctx = (py::capsule)ctx.attr("__capsule__")()) == nullptr)
-      TPY_ERROR_LOC("Invalid context pointer!")
-    ctx_ = Context(c_ctx, false);
-  }
-
 public:
   PyQueryCondition() = delete;
 
@@ -53,11 +40,9 @@ public:
     }
   }
 
-  PyQueryCondition(const string &attribute_name, const string &condition_value,
-                   tiledb_query_condition_op_t op, py::object ctx) {
+  void init(const string &attribute_name, const string &condition_value,
+            tiledb_query_condition_op_t op) {
     try {
-      set_ctx(ctx);
-      qc_ = shared_ptr<QueryCondition>(new QueryCondition(ctx_));
       qc_->init(attribute_name, condition_value, op);
     } catch (TileDBError &e) {
       TPY_ERROR_LOC(e.what());
@@ -65,11 +50,9 @@ public:
   }
 
   template <typename T>
-  PyQueryCondition(const string &attribute_name, T condition_value,
-                   tiledb_query_condition_op_t op, py::object ctx) {
+  void init(const string &attribute_name, T condition_value,
+            tiledb_query_condition_op_t op) {
     try {
-      set_ctx(ctx);
-      qc_ = shared_ptr<QueryCondition>(new QueryCondition(ctx_));
       qc_->init(attribute_name, &condition_value, sizeof(condition_value), op);
     } catch (TileDBError &e) {
       TPY_ERROR_LOC(e.what());
@@ -98,30 +81,77 @@ public:
     return pyqc;
   }
 
+private:
+  PyQueryCondition(shared_ptr<QueryCondition> qc, tiledb_ctx_t *c_ctx)
+      : qc_(qc) {
+    ctx_ = Context(c_ctx, false);
+  }
+
+  void set_ctx(py::object ctx) {
+    tiledb_ctx_t *c_ctx;
+    if ((c_ctx = (py::capsule)ctx.attr("__capsule__")()) == nullptr)
+      TPY_ERROR_LOC("Invalid context pointer!")
+    ctx_ = Context(c_ctx, false);
+  }
 }; // namespace tiledbpy
 
 PYBIND11_MODULE(_query_condition, m) {
   py::class_<PyQueryCondition>(m, "qc")
       .def(py::init<py::object>(), py::arg("ctx") = py::none())
-      .def(py::init<const string &, const string &, tiledb_query_condition_op_t,
-                    py::object>(),
-           py::arg("attribute_name"), py::arg("condition_value"),
-           py::arg("tiledb_query_condition_op_t"), py::arg("ctx") = py::none())
-      .def(py::init<const string &, double, tiledb_query_condition_op_t,
-                    py::object>(),
-           py::arg("attribute_name"), py::arg("condition_value"),
-           py::arg("tiledb_query_condition_op_t"), py::arg("ctx") = py::none())
-      .def(py::init<const string &, float, tiledb_query_condition_op_t,
-                    py::object>(),
-           py::arg("attribute_name"), py::arg("condition_value"),
-           py::arg("tiledb_query_condition_op_t"), py::arg("ctx") = py::none())
-      .def(py::init<const string &, int, tiledb_query_condition_op_t,
-                    py::object>(),
-           py::arg("attribute_name"), py::arg("condition_value"),
-           py::arg("tiledb_query_condition_op_t"), py::arg("ctx") = py::none())
 
-      .def("combine", &PyQueryCondition::combine, py::arg("rhs"),
-           py::arg("combination_op"));
+      /* TODO surely there's a better way to deal with templated PyBind11
+       * functions? but maybe not?
+       * https://github.com/pybind/pybind11/issues/1667
+       */
+
+      .def("init_string",
+           static_cast<void (PyQueryCondition::*)(
+               const string &, const string &, tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+
+      .def("init_uint64",
+           static_cast<void (PyQueryCondition::*)(const string &, uint64_t,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+      .def("init_int64",
+           static_cast<void (PyQueryCondition::*)(const string &, int64_t,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+      .def("init_uint32",
+           static_cast<void (PyQueryCondition::*)(const string &, uint32_t,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+      .def("init_int32",
+           static_cast<void (PyQueryCondition::*)(const string &, int32_t,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+      .def("init_uint16",
+           static_cast<void (PyQueryCondition::*)(const string &, uint16_t,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+      .def("init_int16",
+           static_cast<void (PyQueryCondition::*)(const string &, int16_t,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+      .def("init_uint8",
+           static_cast<void (PyQueryCondition::*)(const string &, uint8_t,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+      .def("init_int8",
+           static_cast<void (PyQueryCondition::*)(const string &, int8_t,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+
+      .def("init_float32",
+           static_cast<void (PyQueryCondition::*)(const string &, float,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+      .def("init_float64",
+           static_cast<void (PyQueryCondition::*)(const string &, double,
+                                                  tiledb_query_condition_op_t)>(
+               &PyQueryCondition::init))
+
+      .def("combine", &PyQueryCondition::combine);
 
   py::enum_<tiledb_query_condition_op_t>(m, "tiledb_query_condition_op_t",
                                          py::arithmetic())
@@ -137,6 +167,32 @@ PYBIND11_MODULE(_query_condition, m) {
       m, "tiledb_query_condition_combination_op_t", py::arithmetic())
       .value("TILEDB_AND", TILEDB_AND)
       .export_values();
+
+  // TODO: DRY, centralize this. It is currently repeated in each translation
+  // unit to avoid uncaught exceptions (typeid does not match)
+  /*
+     We need to make sure C++ TileDBError is translated to a correctly-typed py
+     error. Note that using py::exception(..., "TileDBError") creates a new
+     exception in the *readquery* module, so we must import to reference.
+  */
+  static auto tiledb_py_error =
+      (py::object)py::module::import("tiledb").attr("TileDBError");
+
+  py::register_exception_translator([](std::exception_ptr p) {
+    try {
+      if (p)
+        std::rethrow_exception(p);
+    } catch (const TileDBPyError &e) {
+      PyErr_SetString(tiledb_py_error.ptr(), e.what());
+    } catch (const tiledb::TileDBError &e) {
+      PyErr_SetString(tiledb_py_error.ptr(), e.what());
+    } catch (py::builtin_exception &e) {
+      // just forward the error
+      throw;
+      //} catch (std::runtime_error &e) {
+      //  std::cout << "unexpected runtime_error: " << e.what() << std::endl;
+    }
+  });
 }
 }; // namespace tiledbpy